% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direcetedEdges.R
\name{areNeighborCells}
\alias{areNeighborCells}
\title{Are neighbor cells}
\usage{
areNeighborCells(origin, destination)
}
\arguments{
\item{origin}{vector of origin H3 cell indexes}

\item{destination}{vector of destination H3 cell indexes}
}
\value{
1 if the indexes are neighbors, 0 otherwise.
}
\description{
Returns whether or not the provided H3 cell indexes are neighbors
}
\examples{

areNeighborCells(origin = c("8cbe63562a54bff","8cbe635631103ff"), destination = c("8cbe63562a53bff","8cbe635631003ff"))

}
