% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3r-package.R
\name{cellToBoundary}
\alias{cellToBoundary}
\title{Cell To Boundary}
\usage{
cellToBoundary(cell)
}
\description{
Cell To Boundary
}
\examples{

cellToBoundary(cell = c("8cbe44a96a9c1ff","8c3752c716b43ff"))

}
