% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3lib-package.R
\docType{package}
\name{h3lib-package}
\alias{h3lib}
\alias{h3lib-package}
\title{h3lib}
\description{
h3lib
}
\details{
h3lib exposes the source code of 'H3' (written in 'C') to routines that are callable through 'R'.
Detailed examples on how to call these routines are on the website \url{https://github.com/symbolixau/h3lib#how-do-i-call-these-routines}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/symbolixau/h3lib}
  \item Report bugs at \url{https://github.com/symbolixau/h3lib/issues}
}

}
\author{
\strong{Maintainer}: David Cooley \email{dcooley@symbolix.com.au}

Other contributors:
\itemize{
  \item Ray Shao \email{rshao@symbolix.com.au} [contributor]
}

}
