% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3lib-package.R
\name{latLngToCell}
\alias{latLngToCell}
\title{lat lng to cell}
\usage{
latLngToCell(lat, lng, resolution)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{resolution}{cell resolution}
}
\description{
lat lng to cell
}
\examples{

latLngToCell(lat = c(-37, 37), lng = c(145, -145), resolution = 12L)

}
