% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3r-package.R
\name{latLngToCell}
\alias{latLngToCell}
\title{lat lng to cell}
\usage{
latLngToCell(lat, lng, resolution)
}
\arguments{
\item{lat}{latitude}

\item{lng}{longitude}

\item{resolution}{cell resolution}
}
\description{
lat lng to cell
}
\examples{

latLngToCell(
  lat = c(-37.820197, -37.818476)
  , lng = c(144.983324, -144.967354)
  , resolution = 12L
  )

}
